backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
google-cloud-firestore
google-cloud-storage
firebase-admin
stripe
python-multipart
httpx
reportlab
python-docx
pdfplumber
playwright
